requirements.txt
boto3==1.9.178
python-dateutil==2.8.0
Click==7.0
